requirements.txt
Django==3.0.7
django-solo==1.1.3
RPi.GPIO==0.7.0
skyfield==1.22
